+++ __tests__/unit/env.picker.spec.ts
@@ -21,12 +21,12 @@ describe('EnvPicker', () => {
     });
 
     it('doesn`t set default value for environment', () => {
-      expect(new EnvPicker(undefined as string | undefined, 'test').defaultFor({ production: '123' }).value()).toBe(
-        undefined,
-      );
-      expect(new EnvPicker(undefined as string | undefined, 'test').defaultFor({ development: '123' }).value()).toBe(
-        undefined,
-      );
+      expect(
+        new EnvPicker(undefined as string | undefined, 'test' as string).defaultFor({ production: '123' }).value(),
+      ).toBe(undefined);
+      expect(
+        new EnvPicker(undefined as string | undefined, 'test' as string).defaultFor({ development: '123' }).value(),
+      ).toBe(undefined);
 
       expect(new EnvPicker(false as boolean | undefined, 'test').defaultFor({ test: true }).value()).toBe(false);
       expect(new EnvPicker('', 'test').defaultFor({ test: '123' }).value()).toBe('');
@@ -42,18 +42,18 @@ describe('EnvPicker', () => {
     });
   });
 
-  describe('mapIfExist', () => {
+  describe('mapIfExists', () => {
     it('maps state if it exists', () => {
-      expect(new EnvPicker('123' as string | undefined, 'test').mapIfExist(Number).value()).toBe(123);
-      expect(new EnvPicker(undefined as string | undefined, 'test').default('123').mapIfExist(Number).value()).toBe(
+      expect(new EnvPicker('123' as string | undefined, 'test').mapIfExists(Number).value()).toBe(123);
+      expect(new EnvPicker(undefined as string | undefined, 'test').default('123').mapIfExists(Number).value()).toBe(
         123,
       );
-      expect(new EnvPicker('false', 'test').default('123').mapIfExist(JSON.parse).value()).toBe(false);
+      expect(new EnvPicker('false', 'test').default('123').mapIfExists(JSON.parse).value()).toBe(false);
 
-      expect(new EnvPicker(undefined as string | undefined, 'test').mapIfExist((state) => !!state).value()).toBe(
+      expect(new EnvPicker(undefined as string | undefined, 'test').mapIfExists((state) => !!state).value()).toBe(
         undefined,
       );
-      expect(new EnvPicker(null as string | null, 'test').mapIfExist((state) => !!state).value()).toBe(null);
+      expect(new EnvPicker(null as string | null, 'test').mapIfExists((state) => !!state).value()).toBe(null);
     });
   });